backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
scikit-learn==1.3.2
pandas==2.1.1
numpy==1.26.0
gunicorn==21.2.0